backend/pyproject.toml
[project]
name = "novapay-docs-backend"
version = "1.0.0"
description = "NovaPay Docs Portal - RAG Backend"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "langchain==0.3.14",
    "langchain-openai==0.3.0",
    "langchain-chroma==0.2.2",
    "langchain-community==0.3.14",
    "langchain-core==0.3.30",
    "langsmith==0.2.11",
    "chromadb==0.6.3",
    "python-dotenv==1.0.1",
    "sse-starlette==2.2.1",
]
